feat(tsp): add double bridge perturbation method and fix warnings in instande reader

+++ include/solution.hpp
@@ -6,7 +6,7 @@
 #include "instance.hpp"
 
 struct Solution {
-    std::vector<int> sequence;
+    std::vector<size_t> sequence;
   
     double objective;
